initial, BROKEN, implementation of mock server
I either need to change the server to run on https so Chrome is happy
making requests to it, or I need to change the extension permissions so
it can make requests to http urls.

I think it would be much better to either:

1. setup a nginx https -> http proxy so I can make https requests to the
http backend.
1. create a certificate for a dev instance of mothership. Also use nginx
for this. The benefit here is that we get https/nginx setup for
mothership.