simplify project status calculation
Originally a project's status was based on a somewhat inscrutable calculation that would check the status of all projects. Now, the status is really only about the **master** branch. And it's much simpler.

If the last **master** build succeeds, the status is green. If the last **master** build fails, the status is red. etc.